RF-26: addressing comments
