+++ src/constants/routes.ts
@@ -1,0 +1,7 @@
+const ROUTES = {
+  HOME: 'HomeScreen',
+  PROFILE: 'Profile',
+  SETTINGS: 'SettingsScreen',
+};
+
+export default ROUTES;
